requirements.txt
# Pipeline requirements
# Install in Termux Ubuntu environment:
#   pip install -r requirements.txt --break-system-packages
#
# All packages are pure Python — nothing to compile,
# no Rust toolchain needed.

pyyaml      # config file parsing
requests    # all API calls (Groq, Anthropic, WordPress, Bundle, Buffer)
pytz        # timezone handling for scheduled posts
